requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
httpx==0.25.2
authlib==1.2.1
# AI Chatbot dependencies
openai>=1.6.1,<2.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv==1.0.0
redis==5.0.1
websockets==12.0
aiofiles==23.2.1
# Image processing dependencies
Pillow==10.1.0
